Changed a few things. Click to read what has changed
Changed category to url.
Renamed helpers.py to utils.py. 
Now your pages will be displayed in the navigation! :)

I need to refactor the code, so that the news can be created via the pages module but I still need to figure out how to do it.
And thanks to /u/blueblank for helping me out with the navigation!